+++ src/components/DatePicker/DateRangePicker/index.ts
@@ -1,0 +1,1 @@
+export { DateRangePickerProps, default } from './DateRangePicker';
