+++ src/components/Grid/Row/Row.stories.tsx
@@ -1,5 +1,6 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
+import Col from '../Col';
 import Row, { RowProps } from './Row';
 
 export default {
@@ -7,9 +8,19 @@ export default {
   component: Row,
 } as Meta;
 
-const Template: Story = (args) => <Row {...args} />;
+const Template: Story = (args) => (
+  <Row {...args}>
+    <Col>
+      <div style={{ border: '1px solid black' }}>Hello 1</div>
+    </Col>
+    <Col>
+      <div style={{ border: '1px solid black' }}>Hello 2</div>
+    </Col>
+  </Row>
+);
 
 export const DefaultRow = Template.bind({});
 DefaultRow.args = {
   // props를 넣어주세요.
+  gutter: 16,
 } as RowProps;
