+++ src/components/Menu/Item/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Item';
