+++ src/components/Menu/ItemGroup/ItemGroup.tsx
@@ -1,8 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import React, { MouseEvent, ReactElement, useCallback, useMemo } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { DownOutlined, UpOutlined } from '../Icons';
-import Item from './Item';
+import { DownOutlined, UpOutlined } from '../../Icons';
+import Item from '../Item';
 
 interface ItemGroupProps {
   mode?: 'horizontal' | 'vertical';
