+++ src/components/Menu/ItemGroup/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ItemGroup';
