+++ src/components/Select/Option/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Option';
